+++ tests/test_fst_slice.py
@@ -4046,9 +4046,9 @@ i ; \\
             self.assertEqual('{}', FST('{1}').put_slice(None, norm=False).src)
             self.assertEqual('set()', FST('{1}').put_slice(None, norm='call').src)
 
-            self.assertEqual('{*()}', FST('{1}').put_slice(None, set_norm=True).src)
+            # self.assertEqual('{*()}', FST('{1}').put_slice(None, set_norm=True).src)
             self.assertEqual('{*()}', FST('{1}').put_slice(None, set_norm='star').src)
-            self.assertEqual('{}', FST('{1}').put_slice(None, set_norm=False).src)
+            # self.assertEqual('{}', FST('{1}').put_slice(None, set_norm=False).src)
             self.assertEqual('set()', FST('{1}').put_slice(None, set_norm='call').src)
 
             # put as one
@@ -4088,10 +4088,10 @@ i ; \\
             self.assertEqual('set()', set_.get(0, 0, norm='call').src)
             self.assertEqual('{}', set_.get(0, 0, norm=False).src)
 
-            self.assertEqual('{*()}', set_.get(0, 0, set_norm=True).src)
+            # self.assertEqual('{*()}', set_.get(0, 0, set_norm=True).src)
             self.assertEqual('{*()}', set_.get(0, 0, set_norm='star').src)
             self.assertEqual('set()', set_.get(0, 0, set_norm='call').src)
-            self.assertEqual('{}', set_.get(0, 0, set_norm=False).src)
+            # self.assertEqual('{}', set_.get(0, 0, set_norm=False).src)
 
             # cut empty slice
 
@@ -4105,9 +4105,9 @@ i ; \\
             self.assertEqual(('{1}', '{}'), ((f := FST('{1}')).get_slice(cut=True, norm=False).src, f.src))
             self.assertEqual(('{1}', 'set()'), ((f := FST('{1}')).get_slice(cut=True, norm='call').src, f.src))
 
-            self.assertEqual(('{1}', '{*()}'), ((f := FST('{1}')).get_slice(cut=True, set_norm=True).src, f.src))
+            # self.assertEqual(('{1}', '{*()}'), ((f := FST('{1}')).get_slice(cut=True, set_norm=True).src, f.src))
             self.assertEqual(('{1}', '{*()}'), ((f := FST('{1}')).get_slice(cut=True, set_norm='star').src, f.src))
-            self.assertEqual(('{1}', '{}'), ((f := FST('{1}')).get_slice(cut=True, set_norm=False).src, f.src))
+            # self.assertEqual(('{1}', '{}'), ((f := FST('{1}')).get_slice(cut=True, set_norm=False).src, f.src))
             self.assertEqual(('{1}', 'set()'), ((f := FST('{1}')).get_slice(cut=True, set_norm='call').src, f.src))
 
             # misc
@@ -4243,7 +4243,7 @@ i ; \\
         f = FST(r'''del a, b, c, \
  \
 z''')
-        g = f.get_slice(2, 3, 'targets', cut=True, trivia=(None, 'all-'))
+        g = f.get_slice(2, 3, 'targets', cut=True, trivia=('all-',))
         self.assertEqual('del a, b, \\\nz', f.src)
         self.assertEqual('c,', g.src)
         f.verify()
@@ -4252,7 +4252,7 @@ z''')
         f = FST(r'''del a, b, c, \
  \
 z''')
-        f.put_slice(None, 2, 3, 'targets', trivia=(None, 'all-'))
+        f.put_slice(None, 2, 3, 'targets', trivia=('all-',))
         self.assertEqual('del a, b, \\\nz', f.src)
         f.verify()
 
@@ -4262,7 +4262,7 @@ z''')
         f = FST(r'''a = b = c = \
  \
 z''')
-        g = f.get_slice(2, 3, 'targets', cut=True, trivia=(None, 'all-'))
+        g = f.get_slice(2, 3, 'targets', cut=True, trivia=('all-',))
         self.assertEqual('a = b = \\\nz', f.src)
         self.assertEqual('c =', g.src)
         f.verify()
@@ -4271,7 +4271,7 @@ z''')
         f = FST(r'''a = b = c = \
  \
 z''')
-        f.put_slice(None, 2, 3, 'targets', trivia=(None, 'all-'))
+        f.put_slice(None, 2, 3, 'targets', trivia=('all-',))
         self.assertEqual('a = b = \\\nz', f.src)
         f.verify()